Abstract View Model to separate files
